refactor: dynamo wrappers to it's own modules
