+++ api/environments/dynamodb/wrappers/__init__.py
@@ -1,15 +1,13 @@
-from .types import DynamoProjectMetadata
-from .wrappers import (
-    DynamoEnvironmentAPIKeyWrapper,
+from .environment_api_key_wrapper import DynamoEnvironmentAPIKeyWrapper
+from .environment_wrapper import (
     DynamoEnvironmentV2Wrapper,
     DynamoEnvironmentWrapper,
-    DynamoIdentityWrapper,
 )
+from .identity_wrapper import DynamoIdentityWrapper
 
 __all__ = (
     "DynamoEnvironmentAPIKeyWrapper",
     "DynamoEnvironmentV2Wrapper",
     "DynamoEnvironmentWrapper",
     "DynamoIdentityWrapper",
-    "DynamoProjectMetadata",
 )
